requirement.txt
sentence-transformers==2.2.2
langchain==0.3.1
Flask==2.2.5 
pypdf==3.12.0 
python-dotenv==1.0.1
pinecone[grpc]==5.4.2
langchain-pinecone==0.2.3
langchain-community==0.3.1
langchain-experimental==0.0.1
langchain-google-genai==2.0.11